requirements.txt
django
djangorestframework
openai
pymupdf
python-dotenv
